# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted webmail application API backed by IMAP and SMTP.",
        "title": "Go-Cubemail API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/identities": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns all From: identities for the user. Seeds a default identity from the IMAP username on first access.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "List sender identities",
                "responses": {
                    "200": {
                        "description": "identities list",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Adds a new From: identity (display name, email, reply-to, signature).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Create sender identity",
                "parameters": [
                    {
                        "description": "Identity data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.identityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created identity",
                        "schema": {
                            "$ref": "#/definitions/handler.identityResponse"
                        }
                    },
                    "400": {
                        "description": "email required or invalid body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/identities/{id}": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Replaces all fields of the specified sender identity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update sender identity",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Identity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Identity data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.identityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated identity",
                        "schema": {
                            "$ref": "#/definitions/handler.identityResponse"
                        }
                    },
                    "400": {
                        "description": "invalid id, body, or missing email",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Permanently removes the specified sender identity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Delete sender identity",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Identity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/identities/{id}/default": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Marks the specified identity as the default From: address, clearing the flag on all others.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Set default sender identity",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Identity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/settings": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns the logged-in user's complete UI settings and preferences (e.g. theme, timezone, signatures, out of office status).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get user preferences",
                "responses": {
                    "200": {
                        "description": "Success response containing settings details",
                        "schema": {
                            "$ref": "#/definitions/handler.settingsResponse"
                        }
                    },
                    "500": {
                        "description": "User lookup or database retrieval failure",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Updates the logged-in user's settings. Only provided fields are changed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Save user preferences",
                "parameters": [
                    {
                        "description": "Settings fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.settingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated settings",
                        "schema": {
                            "$ref": "#/definitions/handler.settingsResponse"
                        }
                    },
                    "400": {
                        "description": "invalid body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.identityRequest": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "reply_to": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "handler.identityResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "reply_to": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "handler.settingsRequest": {
            "type": "object",
            "properties": {
                "compose_html": {
                    "type": "boolean"
                },
                "date_format": {
                    "type": "string"
                },
                "oof_enabled": {
                    "type": "boolean"
                },
                "oof_message": {
                    "type": "string"
                },
                "rows_per_page": {
                    "type": "integer"
                },
                "signature_pos": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "handler.settingsResponse": {
            "type": "object",
            "properties": {
                "compose_html": {
                    "type": "boolean"
                },
                "date_format": {
                    "type": "string"
                },
                "oof_enabled": {
                    "type": "boolean"
                },
                "oof_message": {
                    "type": "string"
                },
                "rows_per_page": {
                    "type": "integer"
                },
                "signature_pos": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "The session cookie 'gorc_session' is required for authenticated endpoints.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}